requirements.txt
absl-py==0.10.0
dm-acme==0.1.8
dm-control==0.0.322773188
dm-env==1.2
dm-reverb==0.1.0
dm-tree==0.1.5
gym==0.17.2
tensorflow==2.3.0
